requirements.txt
six==1.16.0
fs==2.4.16
appdirs==1.4.4
opencv-python==4.9.0.80
pillow==10.2.0
aioredis==1.3.1
ujson==5.9.0
redis==4.4.0
sanic==23.12.1
shortuuid==1.0.13
Jinja2==3.0.3
aiofiles~=23.2.1
pydantic==2.5.3
voluptuous==0.14.2
passlib==1.7.4
pickleDB~=0.9.2
json-api==0.1.22
sanic-ext==23.12.0
tortoise-orm==0.20.0
flatten-dict==0.2.0
cachetools==5.3.2
async-caches==0.3.0
repro-zipfile==0.3.1